Added WelchIntegratorThread's functionality
I was encountering a rough bug where the array that this thread pushed to WelchIntegrator was showing up as being filled with all 0s. After some investigation, I realized that ArrayBlockingQueues works with pointers, and not the data values themselves (which makes sense, since threads usually share the same memory space). However, at some point, WelchComputerThread had to overwrite the values in this.periodogram, and if WelchIntegrator wasn't fast enough to catch them, it would get garbage data. To fix this race condition, I added the computations of WelchIntegrator into WelchComputer, since the extra thread wasn't really necessary anyways (it does several orders of magnitude less work than WelchComputer already did), and the alternative was to have WelchComputer block until WelchIntegrator was done, which would slow down the thread that needs to be the fastest.