javadoc style comments
Added javadoc comments. Also created a check to see if this thread is interrupted, to better implement start() and stop() methods in SignalAccumulator. (The interrupted() check avoids a race condition where the user can call stop() and then immediately call start(). Without the check, Queue.done could be set to false before this thread has a chance to close, which would allow the user to create an extra "zombie" thread.)